Add ecs_exim4 module
Add an ecs_exim4 module, which deploys the stephengrier/exim4 container to
the ECS cluster created by the ecs module. This will provision a mail
transport agent (MTA) to act as an MX relay for several domains.

It will deploy two instances of the exim4 container and a target group
added to the NLB created by the ECS module. The ECS services uses the
`host` network type because this provides the real IP address of the
client, whereas with the awsvpc network type the container sees the IP
address of the NLB.

The ECS service sets a max and min percentage of 50/100% of the desired
count so that the number of containers is scaled down then back up
during rolling deployments. This is because the EC2 instances are small
and there isn't enough headroom to have extra containers running.