Add NAT instances to ecs module
The ecs module creates NAT gateways for the ECS hosts to use for
egressing out to the internet. However, NAT gateways are rediculously
expensive. The solution is to use NAT instances instead, which are EC2
instances configured to act as NAT boxes. Because they are just EC2
instances we can use spot pricing to massively reduce the cost.

To be resilient we need a NAT instance in each availability zone. To
ensure this is the case we need an ASG in each AZ which maintains
exactly one spot instance. We also create an ENI for each NAT instance,
which will persist as spot instances come and go and provide a static ID
which can be used in routing tables.