Set the DNS record for imap instance within the imap module
Previously the DNS A record for the imap instance was set in the dns
module with a hard-coded IP address. The imap module knows the IP
address of the elastic IP so change the imap module to take the
route53 zone ID as a parameter and add the record within the module.