Add homeassistant module
Add a homeassistant module that provisions a Home Assistant server on an
EC2 spot instance, as well as a Mosquitto MQTT broker and an Nginx
proxy. These are deployed as docker containers using docker-compose.

The spot instance is part of an autoscaling group with a desired
capacity of 1 to make sure the instance is recreated if the spot
instance is reclaimed by AWS.

This also creates a valid TLS sertificate for Nginx and Mosquitto using
LetsEncrypt. The cert will be recreated each time the instance is
recreated.