Replace aws_ebs_snapshot resources with data resources
The original version of this module created the "cloud1 imap data
snapshot" and "cloud1 ldap data snapshot" EBS snapshots, referencing the
underlying EBS volumes using data resources. However, this has created
the problem that the old EBS volumes can't be deleted because they are
still referenced in the terraform.

Fix this by replacing the aws_ebs_snapshot resources with data resources
that lookup the snapshot IDs. This then allows us to delete the data
resources for the aws_ebs_volumes.

This will cause terraform to want to delete the EBS snapshots in AWS, so
we will need to `terraform state rm` the resources before applying.