Change MX records for grier.org.uk to squarespace
The grier.org.uk domain has been moved from 123-reg.co.uk to
Squarespace.com. Change the MX records to direct email for the domain
through Squarespace's mail servers.

Change the SPF record to include the SPF records for mailgun.org, which
is the mail system used by Squarespace.

Add a mx._domainkey record to make DKIM work correctly. The value for
this was provided in the Squarespace console.