Change cloud init to wait for up to 30 seconds for block devices
The homeassistant instance has failed to provision correctly on two
occasions as a result of the block devices for the EBS volumes not being
present when trying to mount the volumes, even though the cloud init
waits for 5 seconds for them to appear.

We could just wait for longer, but that would slow down the
provisioning, and most of the time it will only take a few seconds for
the devices to appear. So instead, check for the presence of the devices
each second for up to 30 seconds, and break as soon as they are ready.