Add new srv EBS volume for certbot data
The current setup fails to provision occasionally because it requests a
new TLS cert every time the instance is provisioned, and if this happens
often enough it hits Letsencrypt's limit of 5 requests for the same
domain in a rolling 7 day period.

To fix this, create a new EBS volume to persist the Letsencrypt certbot
data, and only request a new certificate in the cloud-init if one
doesn't already exist on the EBS volume. A cron job will be created to
renew the cert when it nears its expiry time, though that will be added
in a subsequent commit.

Mount the new volume on /srv, and store certbot's data under
/srv/certbot. This will allow moving mosquitto's data under
/srv/mosquitto and getting rid of its EBS volume, which will happen in a
later commit.