Make cloud init add a cron job to renew TLS certificate
The `certbot renew` command can be run daily to check whether any
locally managed Let's Encrypt certificates are nearing expiry and renew
them if they are. Add a cron job to run this command to make sure the
home assistant cert if renewed when needed.

If the cert if renewed, the script will also copy the new cert and key
to the mosquitto tls directory and restart nginx and mosquitto services
to pick up the new cert.

Certbot is run inside a docker container, and it's not possible to tell
from the exit code whether it renewed the certificate or not, so use the
--deploy-hook option to write a file to a directory that is mounted from
the host, which can then be used to check whether services need to be
restarted.