Use labels to mount volumes consistently
Device names are unreliable because they change depending on the order
they are attached to the instance, which can result in volumes being
mounted on the wrong mount points. Fix this by adding labels to each
filesystem when they are created, which can then be used to mount them
consistently.